// app/types/user.ts
export type User = {
  name: string;
  first_name: string;
  last_name: string;
  email: string;
  gender: string;
  country: string;
  title: string;
  dateOfBirth: string;
  phone: string;
  country_code: string;
  state: string;
  address: string | null;
  profilePicture: string;
  created_at: string;
};